requirements.txt
# Core Flask Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
Flask-Babel==4.0.0
WTForms==3.1.2
SQLAlchemy==2.0.30
Werkzeug==3.0.1
Jinja2==3.1.3
python-dotenv==1.0.0

# ML and Image Processing (Optimized versions)
tensorflow==2.17.0
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2

# API Integrations
requests==2.32.3
twilio==9.0.1

# Task Queue (Optional - we can skip these for MVP)
# celery==5.4.0
# redis==5.0.4

# Validation and Security
marshmallow==3.21.1
bcrypt==4.1.3

# Development and Testing
pytest==8.2.2

# Additional utilities for our app
email-validator==2.1.1

# Translation support
Babel==2.13.1
polib==1.2.0
